development

+++ Server/smart caterer/src/main/java/com/dineevents/Inventory/DTO/Request/InventoryRequestDTO.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class InventoryRequestDTO {
     private String inventoryName;
-    private int inventoryQuantity;
-    private double unitPrice;
+    private Integer inventoryQuantity;
+    private BigDecimal unitPrice;
 }
