+++ Server/smart caterer/src/main/java/com/dineevents/Inventory/DTO/Response/InventoryResponseDTO.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class InventoryResponseDTO {
     private Long inventoryId;
     private String inventoryName;
-    private int inventoryQuantity;
-    private double unitPrice;
+    private Integer inventoryQuantity;
+    private BigDecimal unitPrice;
 }
